split test modules
